Benchamrks - completed.

+++ src/utils.h
@@ -41,7 +41,7 @@ namespace rdh {
                 return (t_Num >> t_BitPos) & 1;
             }
 
-            constexpr uint8_t SetNthBitToX(uint32_t t_Num, uint8_t t_BitPos, uint8_t t_NewValue)
+            constexpr uint32_t SetNthBitToX(uint32_t t_Num, uint8_t t_BitPos, uint8_t t_NewValue)
             {
                 /* Clear a bit */
                 if (t_NewValue == 0) {
